+++ compose/src/test/java/com/kizitonwose/calendar/compose/WeekCalendarStateTest.kt
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test
 import java.time.DayOfWeek
 import java.time.LocalDate
 
-class WeekCalendarStateTests {
+class WeekCalendarStateTest {
     @Test
     fun `start date update is reflected in the state`() {
         val now = LocalDate.now()
