+++ data/src/test/java/com/kizitonwose/calendar/data/DayOfWeekTest.kt
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import java.time.DayOfWeek
 
-class DayOfWeekTests {
+class DayOfWeekTest {
     @Test
     fun `days until works as expected`() {
         assertEquals(5, DayOfWeek.FRIDAY.daysUntil(DayOfWeek.WEDNESDAY))
