+++ sample/src/androidTest/java/com/kizitonwose/calendar/sample/CalendarViewTest.kt
@@ -41,7 +41,7 @@ import java.time.YearMonth
  */
 @RunWith(AndroidJUnit4::class)
 @LargeTest
-class CalendarViewTests {
+class CalendarViewTest {
     @get:Rule
     val homeScreenRule = ActivityScenarioRule(CalendarViewActivity::class.java)
